Add locking
When a caller knows they want to settle a promise,
but they need to do some work first, e.g. to parse a value,
but they want to skip that work
if some other thread is going to beat them,
then they can lock the future.
A locked future may only have one thread try to settle it.